requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
attrs==24.2.0
blinker==1.9.0
boto3==1.35.90
botocore==1.35.90
build==1.2.2.post1
cachelib==0.9.0
Cartopy==0.24.1
certifi==2024.8.30
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
colorcet==3.1.0
contourpy==1.3.1
cycler==0.12.1
dask==2024.11.2
docopt==0.6.2
eccodes==2.39.1
fastapi==0.115.6
findlibs==0.0.5
Flask==3.1.0
Flask-Caching==2.3.0
fonttools==4.55.0
fsspec==2024.10.0
GDAL==3.10.0
h11==0.14.0
haversine==2.9.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.44.0
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
netCDF4==1.7.2
numba==0.61.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.0.0
pip-tools==7.4.1
pipreqs==0.4.13
platformdirs==4.3.6
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pygrib==2.1.6
pyinterp==2024.11.0
pyopencl==2024.3
pyparsing==3.2.0
pyproj==3.7.0
pyproject_hooks==1.2.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytools==2025.1.1
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.2
redis==5.2.1
requests==2.32.3
rioxarray==0.18.1
s3transfer==0.10.4
scipy==1.14.1
setuptools==75.8.0
shapely==2.0.6
siphash24==1.7
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
xarray==2024.11.0
yarg==0.1.10
zstandard==0.23.0
